+++ datasophon-common/src/main/java/com/datasophon/common/command/VariableCacheCommand.java
@@ -10,9 +10,9 @@ public class VariableCacheCommand implements Serializable{
 
     private String value;
 
-    private Integer clusterId;
+    private Long clusterId;
 
-    public VariableCacheCommand(String key, String value, Integer clusterId) {
+    public VariableCacheCommand(String key, String value, Long clusterId) {
         this.key = key;
         this.value = value;
         this.clusterId = clusterId;
