+++ datasophon-common/src/main/java/com/datasophon/common/vo/ClusterAlertHistoryVO.java
@@ -45,5 +45,5 @@ public record ClusterAlertHistoryVO(
         String createTimeFormatted,
         Date updateTime,
         String updateTimeFormatted,
-        Integer clusterId) implements Serializable {
+        Long clusterId) implements Serializable {
 }
